grid.hexlegend {hexbin}	R Documentation

Add a Legend to a Hexbin Plot

Description

Plots the legend for the plot method of hexbin. Provides a legend indicating the count representations.

Usage

grid.hexlegend(legend, ysize, lcex, inner, style = ,
  	   minarea = 0.05, maxarea = 0.8, mincnt = 1, maxcnt, trans = NULL,
	   inv = NULL, colorcut, density = NULL, border = NULL, pen = NULL,
	   colramp = function(n) { LinGray(n,beg = 90,end = 15) },
           leg.unit = "native")

Arguments

legend
positive number giving width of the legend in inches.

ysize
height of legend in inches

lcex
the characters expansion size for the text in the legend, see par(cex=).

inner
the inner diameter of a hexagon in inches.

style
the hexagon style; see grid.hexagons.

minarea, maxarea
fraction of the cell area for the lowest and largest count, respectively.

mincnt, maxcnt
minimum and maximum count accepted in plot.

trans
a transformation function for the counts such as sqrt.

inv
the inverse transformation function.

colorcut
numeric vector of values covering [0, 1] the determine hexagon color classes boundaries and hexagon legend size boundaries.

border
argument for polygon(). Draw the border for each hexagon.

density
argument for polygon() filling. A 0 causes the polygon not to be filled.

pen
color argument used for polygon(col = .). Determines the color with which the polygon will be filled.

colramp
function accepting an integer n as an argument and returning n colors.

leg.unit
unit to use

Details

The plot method for hexbin objects calls this function to produce a legend by setting the graphics parameters, so hex.legend itself is not a standalone function.

The legend function is preliminary. Later version will include refinements and handle extreme cases (small and large) for cell size and counts.

See the Details section of grid.hexagons's help page.

Value

This function does not return any value.

Author(s)

Dan Carr <dcarr@voxel.galaxy.gmu.edu>

ported by Nicholas Lewin-Koh <kohnicho@comp.nus.edu.sg>

References

see in grid.hexagons.

See Also

hexbin, grid.hexagons, smooth.hexbin, erode.hexbin, hcell2xy, gplot.hexbin,

Examples

## Not a stand alone function; typically only called from plot.hexbin()

## Not run: 
  grid.hexlegend(legend = 2, ysize = 1,lcex=8,inner=0.2,
                 maxcnt = 100, colorcut = c(0.5,0.5))

## End(Not run)

[Package hexbin version 1.27.2 Index]