hcell2xy {hexbin}	R Documentation

Compute X and Y Coordinates for Hexagon Cells

Description

Computes x and y coordinates from hexagon cell id's.

Usage

hcell2xy(hbin, check.erosion = TRUE)

Arguments

hbin
a object of class "hexbin", typically produced by hexbin(*).

check.erosion
logical indicating if only the eroded points should be returned in the case where hbin inherits from "erodebin" (see erodebin-class); is TRUE by default.

Details

The hexbin object hbin contains all the needed information. The purpose of this function is to reduce storage. The cost is additional calculation.

Value

A list with two components of the same length as bin$cell,

x

y

See Also

hexbin.

Examples

x <- rnorm(10000)
y <- rnorm(10000)
plot(x,y, pch=".")
hbin <- hexbin(x,y)
str(xys <- hcell2xy(hbin))
points(xys, cex=1.5, col=2) ; title("hcell2xy( hexbin(..) )", col.main=2)

[Package hexbin version 1.27.2 Index]