hcell2xyInt {hexbin}	R Documentation

Change cell ids to 2d integer coordinate system

Description

Transforms the cell representation of a a lattice into a 2d integer coordinate system.

Usage

  hcell2xyInt(hbin, xbins=NULL, xbnds=NULL, ybnds=NULL, shape=NULL)

Arguments

hbin
a object of class "hexbin", typically produced by hexbin(*).

xbins
the number of bins partitioning the range of xbnds.

xbnds, ybnds
horizontal and vertical limits of the binning region in x or y units respectively; must be numeric vector of length 2.

shape
the shape = yheight/xwidth of the plotting regions.

Details

Takes a grid defined by either the hexbin parameters or dimen in a hexbin object and translates the cell ids for the grid into 2d integer coordinates.

Value

An integer matrix with two columns, i and j representing the integer xy coordinates of the hexagon grid.

i
Integer coordiante of the rows, increases from bottom to top

j
Integer coordiante of the columns, increases from left to right

Author(s)

Nicholas Lewin-Koh

See Also

hcell2xy

Examples

 x<-rnorm(10000)
 y<-rnorm(10000)
 hbin<-hexbin(x,y)
 ijInt<-hcell2xyInt(hbin)

[Package hexbin version 1.27.2 Index]