hexList {hexbin}	R Documentation

Conditional Bivariate Binning into Hexagon Cells

Description

Creates a list of hexbin objects. Basic components are a cell id and a count of points falling in each occupied cell. Basic methods are show(), plot() and summary(), but also erode.

Usage

hexList(x, y = NULL, given = NULL, xbins = 30, shape = 1,
        xbnds = NULL, ybnds = NULL, xlab = NULL, ylab = NULL)

Arguments

x
x coordinate to be binned

y
y coordinate to be binned

given
..

xbins
number of bins partitioning the range of xbnds

shape
the shape = yheight/xwidth of the plotting regions

xbnds
horizontal limits of binning

ybnds
vertical limits of binning

xlab
character strings used as labels for x

ylab
character strings used as labels for y

Details

There is also a coerce method to produce hexbinList objects from lists.

Value

If it is a LIST, use

comp1
Description of 'comp1'

comp2
Description of 'comp2'

...

Author(s)

Nicholas Lewin-Koh

See Also

hexbin, hdiffplot

[Package hexbin version 1.27.2 Index]