hexVP.abline {hexbin}	R Documentation

Add a Straight Line to a HexPlot

Description

This function adds one or more straight lines through the current plot; it is the hexbin version of abline().

Usage

hexVP.abline(hvp, a = NULL, b = NULL, h = numeric(0), v = numeric(0),
             col = "black", lty = 1, lwd = 2, ...)

Arguments

hvp
A hexViewport object that is currently on the active device

a,b
the intercept and slope or if b is NULL, an lm object or a vector of length 2 with c(intercept,slope)

h
the y-value for a horizontal line.

v
the x-value for a vertical line.

col, lty, lwd
line color, type and width.

...
further graphical parameters.

Details

The first form specifies the line in intercept/slope form (alternatively a can be specified on its own and is taken to contain the slope and intercept in vector form).

The h= and v= forms draw horizontal and vertical lines at the specified coordinates.

The coef form specifies the line by a vector containing the slope and intercept.

lm is a regression object which contains reg$coef. If it is of length 1 then the value is taken to be the slope of a line through the origin, otherwise, the first 2 values are taken to be the intercept and slope.

Author(s)

Nicholas Lewin-Koh

See Also

gplot.hexbin, hexViewport, hexMA.loess

[Package hexbin version 1.27.2 Index]