hsmooth-methods {hexbin}	R Documentation

Hexagon Bin Smoothing: Generic hsmooth() and Methods

Description

Methods for the generic function hsmooth in package hexbin: There is currently only the one for hexbin objects.

Usage

## S4 method for signature 'hexbin'
hsmooth(bin, wts)

Arguments

bin
a hexbin object, or an extension such as erodebin-class.

wts
weights vector, see smooth.hexbin

Methods

bin = "hexbin"

is just the smooth.hexbin function (for back compatibility); see its documentation, also for examples.

[Package hexbin version 1.27.2 Index]