inout.hex {hexbin}	R Documentation

Check points for inclusion

Description

Check which points are in hexagons with count <= mincnt.

Usage

inout.hex(hbin, mincnt)

Arguments

hbin
an object of class hexbin.

mincnt
Cutoff, id's for counts less than mincnt are returned

Details

Check which points are in hexagons with count <= mincnt and returns the row ids for those points. One can use the ids to plot low ount hexagons as points instead.

Value

A vector with the row ids of points which fall in hexagons with count less than or equal to mincnt

Author(s)

Nicholas Lewin-Koh

See Also

plotMAhex

[Package hexbin version 1.27.2 Index]