panel.hexboxplot {hexbin}	R Documentation

Boxplot for hexbin lattice plot

Description

A panel function to add a boxplot to a hexbin lattice plot.

Usage

panel.hexboxplot(x, y, xbins = 30,
                 xbnds = c("data", "panel"), ybnds = c("data", "panel"),
                 .prelim = FALSE, .cpl = current.panel.limits(),
                 .xlim = .cpl$xlim, .ylim = .cpl$ylim,
                 .aspect.ratio, type = character(0), cdfcut = 0.25,
                 shadow = 0.05, ..., check.erosion = TRUE)

Arguments

x, y
numeric vector or factor.

xbins
the number of bins partitioning the range of xbnds.

xbnds, ybnds
horizontal and vertical limits of the binning region in x or y units respectively; must be numeric vector of length 2.

.prelim, .cpl, .xlim, .ylim, .aspect.ratio
for internal use.

type
character vector controlling additional augmentation of the display. A "g" in type adds a reference grid, an "hg" adds a hexagonal grid.

cdfcut
number in (0,1) indicating the confidence level for the erosion limits. See erode.hexbin for more information.

shadow
number in (0,1) indicating the confidence level for the erosion limits of a boxplot shadow. See erode.hexbin for more information.

...
potential further arguments passed on.

check.erosion
logical indicating only eroded points should be used for "erodebin" objects; simply passed to hcell2xy, see its documentation.

Value

There is no return value from this function. The results are plotted on the current active device.

Author(s)

Nicholas Lewin-Koh nikko@hailmail.net

See Also

hexbinplot, panel.hexgrid, panel.bwplot

Examples

mixdata <-
    data.frame(x = c(rnorm(5000),rnorm(5000,4,1.5)),
               y = rep(1:2, 5000))
hexbinplot(y ~ x, mixdata, panel = panel.hexboxplot)

[Package hexbin version 1.27.2 Index]