panel.hexgrid {hexbin}	R Documentation

Hexagonal grid for a lattice plot

Description

A panel function to add a hexagonal grid to a lattice plot.

Usage

panel.hexgrid(h, border = grey(0.85))

Arguments

h
an object of class hexbin.

border
a color for the hexagon border colors

Value

There is no return value from this function. The results are plotted on the current active device.

Author(s)

Nicholas Lewin-Koh nikko@hailmail.net

See Also

hexbinplot, hexGraphPaper

[Package hexbin version 1.27.2 Index]