pushHexport {hexbin}	R Documentation

Push a Hexagon Viewport ("hexVP")

Description

Push a Hexagon Viewport ("hexVP", see hexVP-class) on to the tree of (grid) viewports, calling pushViewport.

Usage

pushHexport(hvp, clip = "off")

Arguments

hvp
a hexagon viewport, i.e., an object of class "hexVP", see hexVP-class, typically produced by hexViewport(..).

clip
which viewport to push, either 'on' or 'off' are the allowed arguments, see details.

Details

A hexagon viewport ("hexVP") object has slots for two replicate viewports one with clipping turned on and one with clipping off. This allows toggling the clipping option.

See Also

the underlying pushViewport from the grid package.

[Package hexbin version 1.27.2 Index]