Get coordiantes of the median cell after the erode operation

A method for a eroded hexbin object to extract the coordinates of the median cell. The median is simply the cell with the highest erosion number or the last cell to be eroded.

getHMedian(ebin)

Arguments

ebin

result of erode.hexbin().

Methods

ebin = "erodebin"

...

See also

Examples

set.seed(153)
x <- rnorm(10000)
y <- rnorm(10000)
bin <- hexbin(x,y)

smbin  <- smooth.hexbin(bin)
erodebin <- erode.hexbin(smbin, cdfcut=.5)
getHMedian(erodebin)
#> $x
#> [1] -0.07589057
#> 
#> $y
#> [1] 0.02761399
#>