Loess line for hexbin lattice plot

A panel function to add a loess line to a hexbin lattice plot.

This function contravened CRAN policy and is no longer available.

panel.hexloess(bin, w = NULL, span = 2/3, degree = 1,
               family = c("symmetric", "gaussian"), evaluation = 50, 
               lwd = add.line$lwd, lty = add.line$lty,
               col, col.line = add.line$col, ...)

Arguments

bin

an object of class hexbin.

w

optional counts for object bin.

span

smoothness parameter for loess.

degree

degree of local polynomial used.

family

if "gaussian" fitting is by least-squares, and if "symmetric" a re-descending M-estimator is used.

evaluation

number of points at which to evaluate the smooth curve.

lwd

line weight graphical parameter.

lty

line type graphical parameter.

col

same as col.line.

col.line

line color graphical parameter.

...

optional arguments to loess.control.

Value

There is no return value from this function. The results are plotted on the current active device.

Author

Nicholas Lewin-Koh nikko@hailmail.net

See also